Renamed data folder to dao
