+++ src/dao/fs/ProductsManager.fs.js
@@ -5,7 +5,7 @@ import crypto from "crypto";
 class ProductManager {
     static #products = [];
     constructor() {
-        this.path = "./src/data/fs/files/products.json"
+        this.path = "./src/dao/fs/files/products.json"
         this.init()
     }
