+++ src/middlewares/isValidPassword.mid.js
@@ -1,4 +1,4 @@
-import usersManager from "../data/mongo/managers/userManager.mongo.js";
+import usersManager from "../dao/mongo/managers/userManager.mongo.js";
 
 async function isValidPassword(req, res, next) {
   try {
