+++ src/routers/api/tickets.api.js
@@ -1,5 +1,5 @@
 import CustomRouter from "../customRouter.js";
-import cartsManager from "../../data/mongo/managers/cartsManager.mongo.js";
+import cartsManager from "../../dao/mongo/managers/cartsManager.mongo.js";
 
 class ticketsRouterClass extends CustomRouter {
     init() {
