+++ src/routers/customRouter.js
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { verifyToken } from "../utils/token.util.js";
-import userManagerMongo from "../data/mongo/managers/userManager.mongo.js";
+import userManagerMongo from "../dao/mongo/managers/userManager.mongo.js";
 
 class CustomRouter {
     /*-----------Setup-----------*/
